Fixed alignment of Zone-allocated objects for MIPS and ARM targets.
The 4-byte alignment caused misaligned double-read problems on MIPS in some cases. Not sure why this wasn't discovered on ARM previously.
The alignment has been changed for the simulator as well (using the V8_TARGET_ARCH_XX macros) just in case in the future these objects are used in generated code and the alignment is checked.