Disable mozilla regress-312588.
This test has always been problematic on MIPS boards.
If memory overcommitting is enabled, the test triggers the OOM killer which results in an unexpected crash.
If overcommitting is disabled, the test fails with a proper error message (this is the expected behaviour), but takes too long.

The test was disabled globally in the NewGC merge commit, but that commit was reverted/unmerged on the 3.6 branch.